Change normalization of box matrix elements.
This change in normalization induces a change in the normalization of
the Ktilde matrix to make it more natural. The difference is most easily
seen by comparing

	eq. (2.9) of https://arxiv.org/abs/1802.03100v1	(new convention)
	eq. (46)  of https://arxiv.org/abs/1707.05817v1 (old convention)

In the old normalization, the K matrix depended on m_ref L, i.e. a
finite-volume quantity. Using the new normalization, the K matrix does
not pick up a dependence on any finite-volume quantity, and the
Breit-Wigner parametrization takes its canonical form.